+++ hta/common/trace.py
@@ -91,6 +91,20 @@ class TraceSymbolTable:
     def get_sym_table(self) -> List[str]:
         return self.sym_table
 
+    def add_symbols_to_trace_df(self, trace_df: pd.dataframe, col: str) -> None:
+        """
+        Take a trace dataframe and expand symbols in one of its columns.
+        Args:
+            trace_df (pd.DataFrame): Dataframe for trace from one rank.
+            col (str): column to expand symbols on.
+
+        Returns:
+            None
+        """
+        trace_df[col] = trace_df[col].apply(
+            lambda i: self.sym_table[i] if (i >= 0 and i < len(self.sym_table)) else ""
+        )
+
 
 def parse_trace_dict(trace_file_path: str) -> Dict[str, Any]:
     """
@@ -147,6 +161,7 @@ def compress_df(df: pd.DataFrame) -> Tuple[pd.DataFrame, TraceSymbolTable]:
     args_to_keep = {
         "stream",
         "correlation",
+        "External id",
         "Trace iteration",
         "memory bandwidth (GB/s)",
     }